Move internal versions method to helper class
