+++ eng/update-dependencies/Sync/InternalVersionsHelper.cs
@@ -1,0 +1,56 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Dotnet.Docker.Sync;
+
+internal static class InternalVersionsHelper
+{
+    /// <summary>
+    /// Records the staging pipeline run ID in an easy to parse format. This
+    /// can be used by the sync-internal-release pipeline to record and
+    /// re-apply the same staging builds after resetting the state of the repo
+    /// to match the public release branch.
+    /// </summary>
+    /// <remarks>
+    /// This will only store one staging pipeline run ID per dockerfileVersion
+    /// </remarks>
+    /// <param name="dockerfileVersion">major-minor version</param>
+    /// <param name="stagingPipelineRunId">the build ID of the staging pipeline run</param>
+    public static void RecordInternalVersion(string dockerfileVersion, string stagingPipelineRunId)
+    {
+        const string InternalVersionsFile = "internal-versions.txt";
+
+        // Internal versions file should have one line per dockerfileVersion
+        // Each line should be formatted as: <dockerfileVersion>=<stagingPipelineRunId>
+        //
+        // The preferable way to do this would be to record the version in
+        // manifest.versions.json, however that would require one of the following:
+        // 1) round-trip serialization, which would remove any whitespace/blank lines - which are
+        //    important for keeping the file readable and reducing git merge conflicts
+        // 2) lots of regex JSON manipulation which is error-prone and harder to maintain
+        //
+        // So for now, the separate file and format is a compromise.
+
+        var versionsFilePath = Path.GetFullPath(SpecificCommand.VersionsFilename);
+        var versionsFileDir = Path.GetDirectoryName(versionsFilePath) ?? "";
+        var internalVersionFile = Path.Combine(versionsFileDir, InternalVersionsFile);
+        Dictionary<string, string> versions = [];
+
+        try
+        {
+            // File already exists - read existing versions
+            versions = File.ReadAllLines(internalVersionFile)
+                .Select(line => line.Split('=', 2))
+                .Where(parts => parts.Length == 2)
+                .ToDictionary(parts => parts[0], parts => parts[1]);
+        }
+        catch (FileNotFoundException)
+        {
+            // File doesn't exist - it will be created
+        }
+
+        versions[dockerfileVersion] = stagingPipelineRunId;
+        var versionLines = versions.Select(kv => $"{kv.Key}={kv.Value}");
+        File.WriteAllLines(internalVersionFile, versionLines);
+    }
+}
